refactor: split code into components
