+++ src/routes/safe/components/Settings/SpendingLimit/style.ts
@@ -1,4 +1,4 @@
-import { createStyles } from '@material-ui/core'
+import { createStyles, makeStyles } from '@material-ui/core'
 import {
   background,
   boldFont,
@@ -13,119 +13,121 @@ import {
   xl,
 } from 'src/theme/variables'
 
-export const styles = createStyles({
-  title: {
-    padding: lg,
-    paddingBottom: 0,
-  },
-  hide: {
-    '&:hover': {
-      backgroundColor: '#fff3e2',
+export const useStyles = makeStyles(
+  createStyles({
+    title: {
+      padding: lg,
+      paddingBottom: 0,
     },
-    '&:hover $actions': {
-      visibility: 'initial',
+    hide: {
+      '&:hover': {
+        backgroundColor: '#fff3e2',
+      },
+      '&:hover $actions': {
+        visibility: 'initial',
+      },
     },
-  },
-  actions: {
-    justifyContent: 'flex-end',
-    visibility: 'hidden',
-    minWidth: '100px',
-  },
-  noBorderBottom: {
-    '& > td': {
-      borderBottom: 'none',
+    actions: {
+      justifyContent: 'flex-end',
+      visibility: 'hidden',
+      minWidth: '100px',
     },
-  },
-  annotation: {
-    paddingLeft: lg,
-  },
-  ownersText: {
-    color: secondaryText,
-    '& b': {
-      color: fontColor,
+    noBorderBottom: {
+      '& > td': {
+        borderBottom: 'none',
+      },
     },
-  },
-  container: {
-    padding: lg,
-  },
-  actionButton: {
-    fontWeight: boldFont,
-    marginRight: sm,
-  },
-  buttonRow: {
-    padding: lg,
-    position: 'absolute',
-    left: 0,
-    bottom: 0,
-    boxSizing: 'border-box',
-    width: '100%',
-    justifyContent: 'flex-end',
-    borderTop: `2px solid ${border}`,
-  },
-  modifyBtn: {
-    height: xl,
-    fontSize: smallFontSize,
-  },
-  removeModuleIcon: {
-    marginLeft: lg,
-    cursor: 'pointer',
-  },
-  modalHeading: {
-    boxSizing: 'border-box',
-    justifyContent: 'space-between',
-    maxHeight: '75px',
-    padding: `${sm} ${lg}`,
-  },
-  modalContainer: {
-    minHeight: '369px',
-  },
-  modalManage: {
-    fontSize: lg,
-  },
-  modalClose: {
-    height: '35px',
-    width: '35px',
-  },
-  modalButtonRow: {
-    height: '84px',
-    justifyContent: 'center',
-  },
-  modalButtonRemove: {
-    color: '#fff',
-    backgroundColor: error,
-    height: '42px',
-  },
-  modalName: {
-    textOverflow: 'ellipsis',
-    overflow: 'hidden',
-  },
-  modalUserName: {
-    whiteSpace: 'nowrap',
-  },
-  modalOwner: {
-    backgroundColor: background,
-    padding: md,
-    alignItems: 'center',
-  },
-  modalUser: {
-    justifyContent: 'left',
-  },
-  modalDescription: {
-    padding: md,
-  },
-  modalOpen: {
-    paddingLeft: sm,
-    width: 'auto',
-    '&:hover': {
+    annotation: {
+      paddingLeft: lg,
+    },
+    ownersText: {
+      color: secondaryText,
+      '& b': {
+        color: fontColor,
+      },
+    },
+    container: {
+      padding: lg,
+    },
+    actionButton: {
+      fontWeight: boldFont,
+      marginRight: sm,
+    },
+    buttonRow: {
+      padding: lg,
+      position: 'absolute',
+      left: 0,
+      bottom: 0,
+      boxSizing: 'border-box',
+      width: '100%',
+      justifyContent: 'flex-end',
+      borderTop: `2px solid ${border}`,
+    },
+    modifyBtn: {
+      height: xl,
+      fontSize: smallFontSize,
+    },
+    removeModuleIcon: {
+      marginLeft: lg,
       cursor: 'pointer',
     },
-  },
-  modal: {
-    height: 'auto',
-    maxWidth: 'calc(100% - 30px)',
-    overflow: 'hidden',
-  },
-  amountInput: {
-    width: '100% !important',
-  },
-})
+    modalHeading: {
+      boxSizing: 'border-box',
+      justifyContent: 'space-between',
+      maxHeight: '75px',
+      padding: `${sm} ${lg}`,
+    },
+    modalContainer: {
+      minHeight: '369px',
+    },
+    modalManage: {
+      fontSize: lg,
+    },
+    modalClose: {
+      height: '35px',
+      width: '35px',
+    },
+    modalButtonRow: {
+      height: '84px',
+      justifyContent: 'center',
+    },
+    modalButtonRemove: {
+      color: '#fff',
+      backgroundColor: error,
+      height: '42px',
+    },
+    modalName: {
+      textOverflow: 'ellipsis',
+      overflow: 'hidden',
+    },
+    modalUserName: {
+      whiteSpace: 'nowrap',
+    },
+    modalOwner: {
+      backgroundColor: background,
+      padding: md,
+      alignItems: 'center',
+    },
+    modalUser: {
+      justifyContent: 'left',
+    },
+    modalDescription: {
+      padding: md,
+    },
+    modalOpen: {
+      paddingLeft: sm,
+      width: 'auto',
+      '&:hover': {
+        cursor: 'pointer',
+      },
+    },
+    modal: {
+      height: 'auto',
+      maxWidth: 'calc(100% - 30px)',
+      overflow: 'hidden',
+    },
+    amountInput: {
+      width: '100% !important',
+    },
+  }),
+)
